Remove ZeebeDecoratorBase
